requirements.txt
pypresence # using https://github.com/ghostselfbot/pypresence/archive/master.zip instead cuz it supports custom activity name (waiting for PR to be merged)
flask
python-dotenv